Add data handling, more functional.
Trial constructor is now passed the stimulus list, the stimulus itself,
and the column where stimuli can be found. That felt excessive, so I
added a generate_trials() function to the text.py module (plan to add
one for each type of trial) that takes only the df and colname, and
extracts the stimuli from it, then creates a trial for every stimulus.

run_task() takes an arg referencing the experiment (which can be passed
through the kwargs/template). This allows it to retrieve or modify the
start/end time values or the response. These are then added to the copy
of the row that is part of the trial, and the new row is added to a data
frame. The data frame is returned at the end of the task.