Nearly complete step out of OOP...
Moving from pandas to dicts meant that stimulus lists can be treated as lists, and their members could be passed individually when needed, which more or less gets rid of the need for complex, separate functions for generating trials, tasks, etc.

It also means a set of parameters can be defined early on if needed and passed around, as an alternative to object attributes. So, the `Trial` base class was nearly empty, and there was no need to set parameters, so the derived classes would have just been a redefinition of the `Trial.go()` method. Instead, it makes more sense to not require deriving from the `Trial` class and instead writing a function that runs a trial.

This also allows for more flexibility to create different types of trial, because instead of having a single type of "Text Trial," there can be `visual_ldt_trial()` and `rsvp_trial()`, etc.

Next steps would be to actually write some trial runner functions, and consider removing the `Experiment` class and creating an instance of `pyg.app.EventLoop` in the `run_task()` function or in a higher `run_experiment()` function.